Clients and RabbitMQ and EnergyService working, parent pom.xml created
